requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
yt-dlp==2024.12.23
google-genai==1.65.0
python-dotenv==1.0.1
